refactor: extract ImgGenCore
